Gestion de la vie + Refonte du changement de room avec le current number room

+++ Assets/Scripts/Enemies/Lancier.cs
@@ -53,7 +53,7 @@ public class Lancier : Enemy, IDamagable
         }
     }
 
-    public override void TakeDamage(float amount)
+    public override void TakeDamage(int amount)
     {
         base.TakeDamage(amount);
         if (isAiming)
